OVSDriver: use CFR when checking for overlapping kflows
With the introduction of the meta_id match fields the OVS flow key is no longer
a superset of the OpenFlow match. This broke ind_ovs_kflow_invalidate_overlap,
which translated the kflow key into an of_match and compared it with the
of_match sent by the controller (which may have had the meta_id fields
un-wildcarded).

The fix, which should also be more performant, is to translate the of_match and
kflow key into a CFR.