OVSDriver: periodically synchronize stats
The previous code queried each kflow when OFStateManager requested the stats
for an OpenFlow flow. In the future flows will not be able to efficiently
enumerate their associated kflows, so the stats synchronization needs to
change.

The new code piggybacks on the expiration timer (which should probably be
renamed and converted to use a netlink dump). The kflow keeps track of the
previous stats values, and the diff is added to its parent flow's stats.

This means that stats will be delayed (max ~2.5 seconds). I believe this is
acceptable.