OVSDriver: workaround kernel errno bug
Older versions of the openvswitch kernel module had a bug where they returned
garbage errno values in case of success.

This change ports the workaround already in ivs-ctl to the OVSDriver transact
utility method. It also increases the range of valid errnos based on MAX_ERRNO
in the kernel source.