OVSDriver: compute actions in ind_ovs_kflow_add
The previous code copied the actions used to execute the packet in the upcall
handler into the BH request. This was incorrect because the flowtable could
change between the request being enqueued in an upcall thread and handled in
the main thread.

This change moves the flowtable lookup and action translation into
ind_ovs_kflow_add. This is now symmetric with ind_ovs_kflow_invalidate.