OVSDriver: remove support for OFPP_TABLE
Combining arbitrary packet-out actions with OFPP_TABLE is difficult. Instead of
doing the work to support this in the new multi-table pipeline I'm ripping it
out.

If we want to support OFPP_TABLE in the future we could consider doing it for a
very restricted set of packet-out actions (single output to OFPP_TABLE).