tcam: use bloom_filter module
The assumption is that most lookups in individual shards fail, so we should
make that case faster at the expense of successful lookups. The Bloom filter
uses 2 cache lines (1 per probe) to do a lookup. The hashtable also uses 2
cache lines (bucket head and (in expectation) one node). The difference is the
total number of cachelines used: n/64 for the Bloom filter, and 1.13*n for the
hashtable. Using fewer cachelines makes it more likely that the ones we need
will be already in the cache when we do a lookup.

Reduces time per lookup by 36% in the benchmark. When running benchmarks that
don't fit in L3 cache (200000 flows) this increases to 60%. In practice I
expect the speedup to be somewhere in between because other parts of the switch
(and other processes) are contending for cache space.