emit OVS actions directly from the pipeline
Instead of the standard pipeline concatenating the IVS actions together and
then converting them to OVS actions all at once at the end (in the upcall or
kflow code), translate the actions to OVS actions in the pipeline as we get
them.

The translate_actions.c code is moved to a new file in the pipeline_standard
module.