OVSDriver: use separate netlink sockets for packet-ins
Previously packet-ins were received by an upcall thread and queued to the main
thread. This extra queue was unncessary. Instead, we can create separate
netlink sockets for packet-ins and read them directly in the main thread. We
may also increase the number of packet-in sockets per port in the future, which
will allow better quality of service guarantees.