implement special pipeline handling for in-band management
A new command line option is added to enable in-band management and to
configure the VLAN to use. When a packet with this tag is received on an uplink
it will be detagged and sent to an "inband" internal port. When a packet is
sent from the inband port, it will be tagged with the configured VLAN and sent
out the first live uplink.

Also fixed a bug in the --uplink command line option, and added a missing check
to ind_ovs_uplink_first.