OVSDriver: use a socket to send kflow requests to the main thread
IVS is moving to a multi-process model, so we need to use IPC instead of a
linked list to communicate with the main thread. The advantage of a datagram
socket over a pipe is that the kernel preserves message boundaries.

Each upcall thread now gets its own queue, which improves fairness over the
previous implementation.