OVSDriver: resynchronize vports on multicast socket overrun
When many ports are added to the datapath at once the ind_ovs_multicast_socket
netlink socket buffer can fill up. The next time we read from the socket the
kernel will return an error to let us know this happened. We need to do a dump
to catch any newly added ports.

This was a big problem in mininet, because each instance of IVS uses a separate
datapath in the same network namespace. There's no way to say which datapath
you're interested in when you subscribe to the vport multicast group, so every
switch sees the port-add events from all switches when mininet starts up. The
multicast socket would consistently overrun and some subset of ports would
never get their upcalls handled.