pipeline_lua: add hashtable implementation
This is a Robin Hood hashtable, a variant of open addressed hashtables. It
allocates a single contiguous array of slots. When inserting into the
hashtable, we start probing at an index determined by a hash. We linearly scan
to find an empty slot. The Robin Hood algorithm rearranges entries as it scans
to minimize the average probe length.

The reason we have to build this in to the switch is because it uses the FFI
for performance. We use the FFI to define C structs for the key and value and
to allocate a large array of them. FFI pointers allow raw memory access so we
don't want to expose them to uploaded code. In the future we could create
abstractions on top of the FFI to allow uploaded code to define similar
efficient datastructures.

This hashtable was the last user of the FFI so I've removed access to the FFI
from the sandbox.