megaflow infrastructure
Pipeline implementations are given a mask to fill in when processing a flow.
For every key bit it used when processing a flow, the pipeline must set the
corresponding mask bit to one. This mask is sent to the kernel when inserting a
kflow. Any future packets that agree with the kflow key where the mask bit is
one will hit, reducing the number of kflows that need to be installed.

This change does not yet generate useful masks from the built-in pipelines.