OVSDriver: self test for megaflow masks
In debug builds, this function is run whenever we add or modify a kernel flow.
It double-checks the validity of the mask by randomizing the parts of the key
where the mask is zero. If changing any of those bits causes the output of the
pipeline to change then the mask was incorrect.