workaround GCC bug with optimize pragma
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=58087

This GCC bug causes excessive memory use dependent on the number of functions
covered by the pragma. My theory is that this includes the inline functions in
headers. Moving the pragma after the headers reduces peak compiler memory usage
by 10x, from 2 gigabytes to 200 megabytes. This is important when we're trying
to run 8 compilers in parallel.