OVSDriver: remove ICMPv6 ND from key
The Ubuntu 14.04 kernel has a bug which prevents installing flows which match
on ND fields. The implementation in IVS was also buggy because it was in a
union with the icmpv6 key.

Given these bugs, it's better to remove the ability to match on ND fields and
allow those flows to be installed in the kernel.