OVSDriver: sticky uplinks
The previous algorithm for picking an uplink iterated through all ports to find
the first live one with the uplink bit set. This was more expensive than
necessary, but the main problem is that IFF_RUNNING doesn't mean that the
switch on the other side is ready to forward packets. We don't want to move our
inband OpenFlow connections from a working uplink to a potentially not-working
one.

This change runs the uplink selection algorithm on port events instead of for
every packet. It will try to keep the same uplink as long as it's still live.