packet_trace: replace write loop with single send
AFAIK the only reason write would return a short count is if the socket buffer
filled. In that case doing another write immediately would not be useful.

Unlike write, the send syscall takes a flags argument. We use this to pass
MSG_NOSIGNAL so that the kernel doesn't kill us with SIGPIPE when the other
side disconnects.