OVSDriver: fix a zombie parent death race
PR_SET_DEATHSIG asks the kernel to send an upcall process a SIGKILL when the
parent process dies, but if the parent has already died when we do the prctl
then no signal will be sent. I previously tried to work around this with
kill(parent, 0), which would return ESRCH if the parent had already died. It
turns out this is insufficient because kill returns success if the target is a
zombie.

This change uses the traditional method of watching for the other side of a
pipe to close.