OVSDriver: use libnl port allocation when possible
libnl 3.2.25 broke our method of using the kernel to allocate Netlink ports.
However, in the process it also fixed the underlying issue that necessitated
that hack. This change detects whether we have a newer version of libnl and
uses its port allocation scheme.

The nl_has_capability function is not defined for earlier versions of libnl, so
we declare it as a weak symbol and check at runtime whether it exists.