tests: hitless restart test
This test verifies that an IVS started with the --hitless flag waits until it
receives a bsn_takeover message before revalidating the kernel flowtable.

This is the first non-OFTest here. OFTest isn't designed to do things like
killing and restarting the IVS daemon, so we needed something different. We'll
probably add more tests like this in the future and factor out some of the
netns code into a library.